fix(normalization): Resolve shadowed module export
